backend/requirements.txt
Flask==2.3.2
pymongo==4.6.0
marshmallow==3.20.1
flasgger==0.9.7.1
pytest==7.4.0
python-dotenv==1.0.1
